tegra186-flashtools-native: add bbappend
with an extra patch to l4t_bup_gen.func which completes the
set of boot partitions for tegra186 platforms. This allows us
to use a BUP payload as the source for initializing the boot
partitions from scratch, so we can create an image that
completely reinitalizes a device without using the USB port.