custom-flash-layout: update the jetson-tx2-cboot layout
to set the primary GPT to have the standard 33-sector size.
The default size used in meta-tegra is derived from calculations
in the L4T BSP's flash.sh script, and appear to be both incorrect
(at least in GPT terms) and unnecesary (since its allocating space
for a BCT when it doesn't need to).